requirements.txt
feedparser
PyYAML
requests
beautifulsoup4
markdown
python-dotenv
